reportes_pdf

+++ app/Http/Controllers/Citas/ConsultasMedicasController.php
@@ -6,6 +6,8 @@ use App\User;
 use App\Query;
 use App\Speciality;
 use Illuminate\Http\Request;
+use App\Http\Requests\ValidarAtenderRequest;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
 
 
@@ -23,12 +25,18 @@ class ConsultasMedicasController extends Controller
 
     public function show()
     {
+        if (Auth::user()->hasRole('doctor')) {
        $consultas = Query::join('users as paciente', 'queries.paciente_id', '=', 'paciente.id')
         ->join('users as doctor', 'queries.doctor_id', '=', 'doctor.id')
         ->join('specialities as especialidad', 'queries.speciality_id', '=', 'especialidad.id')
+        ->select(['queries.id as id', 'queries.fecha_inicio', 'paciente.nombres as paciente', 'paciente.apellidos as apellidos', 'doctor.apellidos as doctor', 'especialidad.nombre as especialidad', 'queries.estado as estado'])->where('queries.estado', '=' , 'pendiente')
+        ->where('doctor_id', Auth::user()->id)->get();
+    }else{
+        $consultas = Query::join('users as paciente', 'queries.paciente_id', '=', 'paciente.id')
+        ->join('users as doctor', 'queries.doctor_id', '=', 'doctor.id')
+        ->join('specialities as especialidad', 'queries.speciality_id', '=', 'especialidad.id')
         ->select(['queries.id as id', 'queries.fecha_inicio', 'paciente.nombres as paciente', 'paciente.apellidos as apellidos', 'doctor.apellidos as doctor', 'especialidad.nombre as especialidad', 'queries.estado as estado'])->where('queries.estado', '=' , 'pendiente');
-            
-        //$users = User::select(['id', 'nombres', 'apellidos', 'nacimiento'])->withRole('paciente');
+    }
         return  datatables()->of($consultas)
             ->editColumn('paciente', function ($consulta) {
                 return ucwords($consulta->apellidos. ' ' .$consulta->paciente);
@@ -40,22 +48,37 @@ class ConsultasMedicasController extends Controller
                 return $consulta->getYearsAttribute();
             })
             ->addColumn('action', function ($consulta) {
-                $atender = '<a href="#" onclick="atender('.$consulta->id.')" data-toggle="modal" data-target="#modal_atender" rel="tooltip" title="Atender" class="btn btn-simple btn-primary btn-icon"><i class="material-icons">done_all</i></a>';
+                 $atender ="";
+                 $editar ="";
+                 $eliminar = "";
+                 if (Auth::user()->can('editar-atender')) {
+                   $atender = '<a href="#" onclick="atender('.$consulta->id.')" data-toggle="modal" data-target="#modal_atender" rel="tooltip" title="Atender" class="btn btn-simple btn-primary btn-icon"><i class="material-icons">done_all</i></a>';
+                }
+                if (Auth::user()->can('editar-citas')) {
                 $editar = '<a href="#" onclick="update_cita_pendiente('.$consulta->id.')" data-toggle="modal" data-target="#update_cita_pendiente" rel="tooltip" title="Editar" class="btn btn-simple btn-success btn-icon edit"><i class="material-icons">edit</i></a>';
+                }
+                if (Auth::user()->can('eliminar-citas')) {
                 $eliminar = '<a href="#" onclick="delete_cita_pendiente('.$consulta->id.')" data-toggle="modal" data-target="#eliminar_paciente" rel="tooltip" title="Eliminar" class="btn btn-simple btn-danger btn-icon"><i class="material-icons">close</i></a>';
+                }
 
                 return $atender.$editar.$eliminar;
             })->make(true);
     }
+    
 
     public function carga_atendidos()
     {
+        if (Auth::user()->hasRole('doctor')) {
        $consultas = Query::join('users as paciente', 'queries.paciente_id', '=', 'paciente.id')
         ->join('users as doctor', 'queries.doctor_id', '=', 'doctor.id')
         ->join('specialities as especialidad', 'queries.speciality_id', '=', 'especialidad.id')
+        ->select(['queries.id as id', 'queries.fecha_inicio', 'paciente.nombres as paciente', 'paciente.apellidos as apellidos', 'doctor.apellidos as doctor', 'especialidad.nombre as especialidad'])->where('queries.estado', '=' , 'atendido')->where('doctor_id', Auth::user()->id)->get();
+    }else{
+        $consultas = Query::join('users as paciente', 'queries.paciente_id', '=', 'paciente.id')
+        ->join('users as doctor', 'queries.doctor_id', '=', 'doctor.id')
+        ->join('specialities as especialidad', 'queries.speciality_id', '=', 'especialidad.id')
         ->select(['queries.id as id', 'queries.fecha_inicio', 'paciente.nombres as paciente', 'paciente.apellidos as apellidos', 'doctor.apellidos as doctor', 'especialidad.nombre as especialidad'])->where('queries.estado', '=' , 'atendido');
-            
-        //$users = User::select(['id', 'nombres', 'apellidos', 'nacimiento'])->withRole('paciente');
+    }
         return  datatables()->of($consultas)
             ->editColumn('paciente', function ($consulta) {
                 return ucwords($consulta->apellidos. ' ' .$consulta->paciente);
@@ -67,20 +90,31 @@ class ConsultasMedicasController extends Controller
                 return $consulta->getYearsAttribute();
             })
             ->addColumn('action', function ($consulta) {
+                $ver ="";
+                $editar ="";
+                $eliminar ="";
+                if (Auth::user()->can('leer-citas')) {
+                $ver = '<a href="#" onclick="ver_cita('.$consulta->id.')" data-toggle="modal" data-target="#modal_ver" rel="tooltip" title="Ver consulta" class="btn btn-simple btn-primary btn-icon edit"><i class="material-icons">remove_red_eye</i></a>';
+            }
+            if (Auth::user()->can('editar-atender')) {
                 $editar = '<a href="#" onclick="atender('.$consulta->id.')" data-toggle="modal" data-target="#modal_atender" rel="tooltip" title="Editar" class="btn btn-simple btn-success btn-icon edit"><i class="material-icons">edit</i></a>';
+            }
+            if (Auth::user()->can('eliminar-consultas')) {
                 $eliminar = '<a href="#" onclick="delete_cita_pendiente('.$consulta->id.')" data-toggle="modal" data-target="#eliminar_paciente" rel="tooltip" title="Eliminar" class="btn btn-simple btn-danger btn-icon"><i class="material-icons">close</i></a>';
-                return $editar.$eliminar;
+            }
+                return $ver.$editar.$eliminar;
             })->make(true);
     }
+    
 
-    public function atender(Query $queries, User $users, $id)
+    public function atender(ValidarAtenderRequest $request, Query $queries, User $users, $id)
     {
         $atender =   $queries->findOrFail($id);
         $paciente =  $users->findOrFail($atender->paciente_id);
         $visitas =   $queries->all()->where('paciente_id', '=', $atender->paciente_id)->where('estado', '=', 'atendido')->count();
         return response()->json([
                 'success' => true,
-                "paciente" => $paciente->nombres . ' '. $paciente->apellidos,
+                "paciente"=> $paciente->nombres . ' '. $paciente->apellidos,
                 "edad"    => $paciente->getYearsAttribute(),
                 "visitas" => $visitas,
                 "id"      => $atender->id
@@ -106,7 +140,7 @@ class ConsultasMedicasController extends Controller
         ]);
     }
 
-    public function update(Query $queries,Request $request,$id)
+    public function update(ValidarAtenderRequest $request, Query $queries,$id)
     {
         if($request->ajax()){
             $consulta = $queries->findOrFail($id);
